vk3/requirements.txt
pandas
playwright
openpyxl
python-docx
pillow
requests